refactor(binding-behavior): shorten names
