+++ packages/runtime/src/resources/binding-behaviors/self.ts
